Add x0.25 and x0.5 slow-speed options to CPU speed menu
Change cpu_power from int to float to support fractional speed
multipliers. Updates timing, audio ratio, and status bar display
accordingly.

+++ src/config.h
@@ -124,7 +124,7 @@ typedef struct {
 		bool baud_high[USE_TAPE_TMP];
 	#endif
 	bool compress_state;
-	int cpu_power;
+	float cpu_power;
 	bool full_speed, drive_vm_in_opecode;
 	
 	// recent files
